+++ lib/shark/consent_service.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require 'shark/consent_service/base'
-require 'shark/consent_service/consent'
-
-module Shark
-  module ConsentService
-  end
-end